fixes try catch and converts more test files to typescript
